[config] Move getAccountUsername to CLI package to reduce usage
