+++ packages/@expo/cli/src/utils/downloadExpoGoAsync.ts
@@ -1,4 +1,3 @@
-import { getExpoHomeDirectory } from '@expo/config/build/getUserState';
 import path from 'path';
 import ProgressBar from 'progress';
 import { gt } from 'semver';
@@ -9,6 +8,7 @@ import { ora } from './ora';
 import { profile } from './profile';
 import { createProgressBar } from './progress';
 import { getVersionsAsync, SDKVersion } from '../api/getVersions';
+import { getExpoHomeDirectory } from '../api/user/UserSettings';
 import { Log } from '../log';
 
 const debug = require('debug')('expo:utils:downloadExpoGo') as typeof console.log;
