+++ packages/@expo/config/src/index.ts
@@ -4,4 +4,3 @@ export * from './Config';
 export * from './Config.types';
 export * from './getExpoSDKVersion';
 export * from './Errors';
-export { getAccountUsername } from './getAccountUsername';
